Rudementary simulation

+++ src/Common/Models/Blob.cs
@@ -5,7 +5,7 @@ public record Blob(Position Position, Velocity Velocity, int Diameter, Color Col
 }
 
 public record Position(int X, int Y);
-public record Velocity(int X, int Y);
+public record Velocity(double X, double Y);
 
 public record Color(byte R, byte G, byte B) {
 	public Color(Random random) : this((byte)random.Next(0, 255), (byte)random.Next(0, 255), (byte)random.Next(0, 255)) { }
